We need to start making changelogs

+++ src/main/java/net/minecraft/client/gui/Gui.java
@@ -47,18 +47,18 @@ public class Gui
     /**
      * Draws a solid color rectangle with the specified coordinates and color (ARGB format). Args: x1, y1, x2, y2, color
      */
-    public static void drawRect(float left, float top, float right, float bottom, int color)
+    public static void drawRect(double left, double top, double right, double bottom, int color)
     {
         if (left < right)
         {
-            float i = left;
+            double i = left;
             left = right;
             right = i;
         }
 
         if (top < bottom)
         {
-            float j = top;
+            double j = top;
             top = bottom;
             bottom = j;
         }
